+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/avatar/Avatar.kt
@@ -25,6 +25,7 @@ import androidx.compose.ui.platform.LocalInspectionMode
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
 import coil.annotation.ExperimentalCoilApi
 import coil.compose.ImagePainter
@@ -43,7 +44,7 @@ import io.getstream.chat.android.compose.ui.theme.ChatTheme
  * @param textStyle The text style of the [initials] text.
  * @param placeholderPainter The placeholder to render while loading is in progress.
  * @param contentDescription Description of the image.
- * @param initialsAvatarOffset The offset for the [InitialsAvatar] depending on the avatar position.
+ * @param initialsAvatarOffset The initials offset to apply to the avatar.
  * @param onClick OnClick action, that can be nullable.
  */
 @ExperimentalCoilApi
@@ -56,7 +57,7 @@ public fun Avatar(
     textStyle: TextStyle = ChatTheme.typography.title3Bold,
     placeholderPainter: Painter? = null,
     contentDescription: String? = null,
-    initialsAvatarOffset: InitialsAvatarOffset = InitialsAvatarOffset.Center,
+    initialsAvatarOffset: DpOffset = DpOffset(0.dp, 0.dp),
     onClick: (() -> Unit)? = null,
 ) {
     if (LocalInspectionMode.current && imageUrl.isNotBlank()) {