+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/avatar/GroupAvatar.kt
@@ -32,7 +32,7 @@ import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.text.TextStyle
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
-import io.getstream.chat.android.compose.ui.util.extensions.internal.getAvatarPosition
+import io.getstream.chat.android.compose.ui.util.extensions.internal.getAvatarPositionOffset
 
 /**
  * Default max number of avatars shown in the grid.
@@ -85,7 +85,11 @@ public fun GroupAvatar(
                         shape = RectangleShape,
                         textStyle = textStyle,
                         showOnlineIndicator = false,
-                        initialsAvatarOffset = imageIndex.getAvatarPosition(imageCount)
+                        initialsAvatarOffset = getAvatarPositionOffset(
+                            dimens = ChatTheme.dimens,
+                            userPosition = imageIndex,
+                            memberCount = imageCount
+                        )
                     )
                 }
             }
@@ -106,7 +110,11 @@ public fun GroupAvatar(
                         shape = RectangleShape,
                         textStyle = textStyle,
                         showOnlineIndicator = false,
-                        initialsAvatarOffset = imageIndex.getAvatarPosition(imageCount)
+                        initialsAvatarOffset = getAvatarPositionOffset(
+                            dimens = ChatTheme.dimens,
+                            userPosition = imageIndex,
+                            memberCount = imageCount
+                        )
                     )
                 }
             }